+++ src/Observers/VariantSuperAttributeObserver.php
@@ -15,6 +15,7 @@
 namespace TechDivision\Import\Product\Variant\Observers;
 
 use TechDivision\Import\Dbal\Utils\EntityStatus;
+use TechDivision\Import\Utils\RegistryKeys;
 use TechDivision\Import\Utils\StoreViewCodes;
 use TechDivision\Import\Utils\BackendTypeKeys;
 use TechDivision\Import\Observers\StateDetectorInterface;
@@ -139,7 +140,6 @@ class VariantSuperAttributeObserver extends AbstractProductImportObserver implem
      */
     protected function process()
     {
-
         // extract the parent SKU and attribute code from the row
         $parentSku = $this->getValue(ColumnKeys::VARIANT_PARENT_SKU);
         $attributeCode = $this->getValue(ColumnKeys::VARIANT_ATTRIBUTE_CODE);
@@ -179,10 +179,21 @@ class VariantSuperAttributeObserver extends AbstractProductImportObserver implem
                 new \Exception($message, null, $e)
             );
 
-            // query whether or not, debug mode is enabled
-            if ($this->isDebugMode()) {
+            // Query whether strict mode is disabled
+            if (!$this->isStrictMode()) {
                 // log a warning and return immediately
                 $this->getSystemLogger()->warning($wrappedException->getMessage());
+                $this->mergeStatus(
+                    array(
+                        RegistryKeys::NO_STRICT_VALIDATIONS => array(
+                            basename($this->getFilename()) => array(
+                                $this->getLineNumber() => array(
+                                    ColumnKeys::VARIANT_ATTRIBUTE_CODE =>  $wrappedException->getMessage()
+                                )
+                            )
+                        )
+                    )
+                );
                 return;
             }
 
@@ -222,9 +233,20 @@ class VariantSuperAttributeObserver extends AbstractProductImportObserver implem
             );
 
             // query whether or not, debug mode is enabled
-            if ($this->isDebugMode()) {
+            if (!$this->isStrictMode()) {
                 // log a warning and return immediately
                 $this->getSystemLogger()->warning($wrappedException->getMessage());
+                $this->mergeStatus(
+                    array(
+                        RegistryKeys::NO_STRICT_VALIDATIONS => array(
+                            basename($this->getFilename()) => array(
+                                $this->getLineNumber() => array(
+                                    ColumnKeys::VARIANT_PARENT_SKU =>  $wrappedException->getMessage()
+                                )
+                            )
+                        )
+                    )
+                );
                 return;
             }
 
@@ -277,7 +299,6 @@ class VariantSuperAttributeObserver extends AbstractProductImportObserver implem
      */
     protected function prepareProductSuperAttributeAttributes()
     {
-
         // load the parent ID
         $parentId = $this->getParentId();